Add createBookData and update deleteBook functions
Made a createBookData function to produce data for data-book
attribute and is used for easy reference on DOM elements.
Use data-book as reference for deleting the DOM Book Cards so that a
change in Card DOM structure won't affect the function.
Updated the function to delete the book from the myLibrary.